refactor: test 클래스 이름, 패키지 구조 변경
