+++ src/test/java/techcourse/myblog/web/IndexControllerTest.java
@@ -4,10 +4,12 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
+import techcourse.myblog.web.test.WebClientGenerator;
+
 import static org.springframework.http.HttpMethod.GET;
 
 @ExtendWith(SpringExtension.class)
-class IndexControllerTests extends WebClientGenerator {
+class IndexControllerTest extends WebClientGenerator {
 
     @Test
     void index() {
